+++ main.go
@@ -8,19 +8,15 @@ import (
 
 func main() {
 
-
-
-
-
 	e := echo.New()
-	//e.GET("/", controller.Proto)
+
 	e.POST("/client", controller.Options)
-	e.GET("/summary", controller.Summary)
+
 
 	e.Use(middleware.CORSWithConfig(middleware.CORSConfig{
 		AllowOrigins: []string{"*"},
 		AllowMethods: []string{echo.GET, echo.PUT, echo.PATCH, echo.POST, echo.DELETE},
 	}))
-	//e.GET("/j", controller.JSON)
+
 	e.Logger.Fatal(e.Start(":8080"))
 }
